requirements.txt
Django==4.0.4
celery==5.2.6
redis==4.2.2
python-magic==0.4.25
mozilla-django-oidc==2.0.0
djangorestframework==3.13.1
